fix(ArtyHub-Mislata/sprints#3): Cambiado el double por el Bigdecimal

+++ src/main/java/es/artyhub/tienda_back/domain/dto/DetailDto.java
@@ -3,6 +3,7 @@ package es.artyhub.tienda_back.domain.dto;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
+import java.math.BigDecimal;
 
 public record DetailDto(
         Long id,
@@ -13,5 +14,5 @@ public record DetailDto(
 
         @NotBlank(message = "El precio no puede ser vacío")
         @Positive(message = "El precio debe ser mayor a 0")
-        double price) {
+        BigDecimal price) {
 }
